Fixes for serialization and reflection features
